requirements.txt
argcomplete==3.1.1
click==8.1.4
colorama==0.4.6
contourpy==1.0.6
customtkinter==5.2.0
cycler==0.11.0
darkdetect==0.8.0
et-xmlfile==1.1.0
fonttools==4.38.0
keras==2.13.1
kiwisolver==1.4.4
matplotlib==3.6.2
mysql-connector-python==8.0.31
numpy==1.23.5
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pipx==1.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
userpath==1.8.0
